requirements.txt
aiocache==0.11.1
aiohttp==3.6.2
cachetools==4.1.0
discord.py==1.5.1
matplotlib==3.3.3
pymongo==3.10.1
python-dotenv==0.15.0
